Fix lineardb3 read/write ordering on windows
Windows server builds were experiencing map errors and inconsistencies.
kissTest failed on windows. The cause appears to be read/write ordering
as identified in fopen documentation

> For files open for update (those which include a "+" sign), on which
> both input and output operations are allowed, the stream shall be
> flushed (fflush) or repositioned (fseek, fsetpos, rewind) before a
> reading operation that follows a writing operation. The stream shall be
> repositioned (fseek, fsetpos, rewind) before a writing operation that
> follows a reading operation (whenever that operation did not reach the
> end-of-file).

This patch adds a conditional wrapper around lineardb3 file operations to
add a seek when necessary. It is enabled by default on windows.
(Presumably the continued operation of the main servers means this is
not an issue on Linux due to their implementation of the undefined
behavior.)