Assign the person that named the baby as leader
This makes babies follow the person that named them.
Naming happens only once at most, and sometimes not at all.

Most of the time mothers name their children. But sometimes
children are taken care of by their aunts, uncles, or strangers.

Naming a child takes a bit of effort and proves that
the person is at least paying attention to the child.