Fix for storage tank to work offscreen
Server keeps track of certain movement trans, the NSWE moving ones. For movement from position A to position B, server passes the next decay trans of position B to check if it is eligible to be "always live tracked". So e.g. when water moves from pipe A to pipe B, the transition "-1 + pipe with water = -1 + moving water" is passed. So the decay of pipe B is kept alive, and water will flow onward to pipe C etc. But the server doesnt check if position A's next decay is eligible. Thats the fix here, check for position A as well.