2hol's adaptation of the craving mechanics
Instead of using the # of generation of a lineage, we use the maximum depth of food ever eaten in the lineage to gauge how advance a lineage is. The next craving is drawn from a pool with certain configurable probability to "advance the tech" (i.e. craving food higher up on the tech tree than all the food ever eaten)